Extract resolvers and mutations for shop
